+++ app/src/main/java/qds/puck/App.kt
@@ -1,7 +1,7 @@
 package qds.puck
 
 import androidx.compose.runtime.Composable
-import qds.puck.ui.login.LoginScreen
+import qds.puck.login.LoginScreen
 
 @Composable
 fun App() {
